🔧 fix: 修复 Windows x64 指针截断与类型转换警告
- loopback_dev/ethernet_dev/tun_dev: (long)ptr → (intptr_t)ptr
- arp.h: arp_hdr_len 返回类型 unsigned long → size_t

+++ src/ehip-netdev-class/ethernet_dev.c
@@ -49,10 +49,10 @@ const struct ehip_netdev_trait_ops ethernet_dev_trait_ops = {
 
 int ethernet_dev_trait_change(ehip_netdev_t *netdev, const void *type_ptr, const void *src_ptr){
     struct ethernet_trait *netdev_ethernet_trait = (struct ethernet_trait *)ehip_netdev_to_trait(netdev);
-    long long_offset = (long)type_ptr-(long)netdev;
+    intptr_t long_offset = (intptr_t)type_ptr-(intptr_t)netdev;
     uint16_t offset;
     int ret = 0;
-    if(long_offset >= (long)(sizeof(ehip_netdev_t) + sizeof(struct ethernet_trait))){
+    if(long_offset >= (intptr_t)(sizeof(ehip_netdev_t) + sizeof(struct ethernet_trait))){
         return EH_RET_INVALID_PARAM;
     }
     offset = (uint16_t)long_offset;
